requirements.txt
anthropic
python-docx
python-dotenv
PyMuPDF